15/03/2025:18/51 work on transactions

+++ database/migrations/2025_03_15_173614_update_wallet_transactions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
           
         Schema::table('wallet_transactions', function (Blueprint $table) {
             
-            $table->integer('payment_id')->unsigned();
+            $table->bigInteger('payment_id')->unsigned();
             $table->foreign('payment_id')->references('id')->on('payments')->onDelete('cascade')->onUpdate('cascade');
               
             // Drop the existing foreign key constraint
